// src/domains/service-clients/payment/proto/generated/payment_service.ts
export interface CreatePaymentSuccess {
  provider: Provider;
  paymentId: string;
  status: string;
  orderId: string;
  stripe?: StripePaymentSession | undefined;
  razorpay?: RazorpayPaymentSession | undefined;
  paypal?: PayPalPaymentSession | undefined;
}